+++ src/HSquared.jl
@@ -1,5 +1,6 @@
 module HSquared
 
+using LinearAlgebra
 using SparseArrays
 
 export AutoBackend,
@@ -7,11 +8,13 @@ export AutoBackend,
     CPUBackend,
     CUDABackend,
     GaussianFamily,
+    GaussianLikelihoodResult,
     HSControl,
     Pedigree,
     Phase0NotImplementedError,
     animal_model_spec,
     fit_animal_model,
+    gaussian_loglik,
     hsquared,
     inbreeding_coefficients,
     normalize_pedigree,
@@ -22,6 +25,7 @@ include("control.jl")
 include("errors.jl")
 include("pedigree.jl")
 include("model_spec.jl")
+include("likelihood.jl")
 include("placeholders.jl")
 
 end